I have refactored the Board logic and Mineplacer Logic.
Button positions in the board can be easily tracked by a class extending button, and having a row and column property. Then an array of type of the same class is used to store those button objects created.

Used recursion to achieve mine checking.

I also added the resetGame Functionality.